refac: A little code clean up
