+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.Infrastructure/CalculationResultClient.cs
@@ -17,7 +17,6 @@ using System.Net.Http.Headers;
 using System.Net.Http.Json;
 using Energinet.DataHub.Core.JsonSerialization;
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces;
-// TODO: Should we avoid referencing the DatabricksClient project "just" to get access to the DatabricksOptions?
 using Energinet.DataHub.Wholesale.Components.DatabricksClient;
 using Energinet.DataHub.Wholesale.Domain.ProcessStepResultAggregate;
 using Energinet.DataHub.Wholesale.Infrastructure.Processes;
@@ -62,8 +61,6 @@ public class CalculationResultClient : ICalculationResultClient
         };
         var requestString = _jsonSerializer.Serialize(requestObject);
 
-        // TODO: Use databricks workspace url from config
-        // TODO: Enable SQL statement execution API in terraform: https://registry.terraform.io/providers/databricks/databricks/latest/docs/data-sources/sql_warehouse#attribute-reference
         var response = await _httpClient.PostAsJsonAsync(StatementsEndpointPath, new StringContent(requestString)).ConfigureAwait(false);
 
         var jsonResponse = response.Content.ReadAsStringAsync().ConfigureAwait(false).GetAwaiter().GetResult();
@@ -83,6 +80,7 @@ public class CalculationResultClient : ICalculationResultClient
         httpClient.DefaultRequestHeaders.Accept.Clear();
         httpClient.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
         httpClient.DefaultRequestHeaders.TryAddWithoutValidation("Content-Type", "application/json");
+        httpClient.BaseAddress = new Uri(options.Value.DATABRICKS_WORKSPACE_URL);
     }
 
     // TODO: Unit test the SQL (ensure it works as expected)
