+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.Infrastructure/ProcessResultPointFactory.cs
@@ -24,7 +24,6 @@ public class ProcessResultPointFactory : IProcessResultPointFactory
      dynamic jsonObject = JsonConvert.DeserializeObject(input) ?? throw new InvalidOperationException();
      var result = jsonObject.result.data_array;
 
-     // TODO: Why is the list not used?
      var list = new List<ProcessResultPoint>();
      foreach (var res in result)
      {
@@ -32,6 +31,6 @@ public class ProcessResultPointFactory : IProcessResultPointFactory
          list.Add(new ProcessResultPoint(res[1], res[2], res[0]));
      }
 
-     return result;
+     return list;
  }
 }
