+++ source/dotnet/wholesale-api/DatabricksClient/DatabricksWheelClient/JobWheelTask.cs
@@ -12,12 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Microsoft.Azure.Databricks.Client;
+using Newtonsoft.Json;
 
-namespace Energinet.DataHub.Wholesale.Components.DatabricksClient
+namespace Energinet.DataHub.Wholesale.Components.DatabricksClient.DatabricksWheelClient;
+
+public class JobWheelTask
 {
-    public interface IJobsWheelApi : IJobsApi
-    {
-        Task<WheelJob> GetWheel(long jobId, CancellationToken cancellationToken = default);
-    }
+    [JsonProperty(PropertyName = "spark_python_task")]
+    public SparkPythonTask SparkPythonTask { get; set; }
+
+    [JsonProperty(PropertyName = "python_wheel_task")]
+    public PythonWheelTask PythonWheelTask { get; set; }
 }
