+++ source/dotnet/wholesale-api/DatabricksClient/DatabricksWheelClient/WheelRunSettings.cs
@@ -14,7 +14,7 @@
 
 using Newtonsoft.Json;
 
-namespace Energinet.DataHub.Wholesale.Components.DatabricksClient
+namespace Energinet.DataHub.Wholesale.Components.DatabricksClient.DatabricksWheelClient
 {
     public class WheelRunSettings
     {
