+++ source/dotnet/wholesale-api/Infrastructure/Calculations/DatabricksCalculatorJobSelector.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Components.DatabricksClient;
+using Energinet.DataHub.Wholesale.Components.DatabricksClient.DatabricksWheelClient;
 
 namespace Energinet.DataHub.Wholesale.Infrastructure.Calculations;
